+++ src/main/java/com/easyfood/EasyFoodApplication/Exception/DailyFoodNotFoundException.java
@@ -1,9 +1,9 @@
 package com.easyfood.EasyFoodApplication.Exception;
 
 public class DailyFoodNotFoundException extends Exception{
-        private int id;
+        private long id;
 
-        public DailyFoodNotFoundException(int id){
+        public DailyFoodNotFoundException(long id){
             super(String.format("Book is not found with id : '%s'", id));
         }
